fix: student dashboard layout
